working on forest gen

+++ dogue/src/com/deweyvm/dogue/ui/Panel.scala
@@ -1,6 +1,6 @@
 package com.deweyvm.dogue.ui
 
-import com.deweyvm.gleany.data.{Point2f, Recti}
+import com.deweyvm.gleany.data.{Point2d, Point2f, Recti}
 import com.deweyvm.dogue.graphics.{Renderer, RectSprite}
 import com.deweyvm.gleany.graphics.Color
 import com.deweyvm.dogue.entities.Tile
@@ -9,7 +9,7 @@ import com.deweyvm.dogue.Dogue
 
 class Panel(val x:Int, val y:Int, val width:Int, val height:Int, bgColor:Color) {
   def contains(i:Int, j:Int):Boolean =
-    getRects exists { _.contains(Point2f(i,j)) }
+    getRects exists { _.contains(Point2d(i,j)) }
 
   def getRects:Vector[Recti] = Vector(Recti(x, y, width, height))
   def update:Panel = this
